+++ src/shaderlib/project/project.glsl.js
@@ -27,7 +27,6 @@ const float WORLD_SCALE = TILE_SIZE / (PI * 2.0);
 const float PROJECT_IDENTITY = 0.;
 const float PROJECT_MERCATOR = 1.;
 const float PROJECT_MERCATOR_OFFSETS = 2.;
-const float PROJECT_UTM_OFFSETS = 4.;
 
 uniform float projectionMode;
 uniform float projectionScale;
@@ -48,11 +47,6 @@ float project_scale(float meters) {
 }
 
 vec2 project_scale(vec2 meters) {
-  if (projectionMode == PROJECT_UTM_OFFSETS) {
-    // In UTM projection mode, latitude and longitude are no longer independent
-    // in the scaling transformation. The scaler is therefore a mat2.
-    return meters * projectionPixelsPerUnitUTM;
-  }
   return meters * projectionPixelsPerUnit.xy;
 }
 
@@ -108,7 +102,7 @@ vec2 project_position(vec2 position) {
 }
 
 vec4 project_to_clipspace(vec4 position) {
-  if (projectionMode == PROJECT_MERCATOR_OFFSETS || projectionMode == PROJECT_UTM_OFFSETS) {
+  if (projectionMode == PROJECT_MERCATOR_OFFSETS) {
     position.w *= projectionPixelsPerUnit.z;
   }
   return projectionMatrix * position + projectionCenter;